+++ inluppar/inlupp1/linked_list.h
@@ -1,6 +1,7 @@
 #pragma once
 #include <stdbool.h>
 #include "hash_table.h"
+#include <stddef.h>   
 
 typedef struct list ioopm_list_t; /// Meta: struct definition goes in C file
 typedef struct link link_t;
@@ -32,7 +33,7 @@ void ioopm_linked_list_prepend(ioopm_list_t *list, int value);
 /// @param list the linked list that will be extended
 /// @param index the position in the list
 /// @param value the value to be inserted 
-void ioopm_linked_list_insert(ioopm_list_t *list, int index, int value);
+void ioopm_linked_list_insert(ioopm_list_t *list, size_t index, int value);
 
 /// @brief Remove an element from a linked list in O(n) time.
 /// The valid values of index are [0,n-1] for a list of n elements,
@@ -40,7 +41,7 @@ void ioopm_linked_list_insert(ioopm_list_t *list, int index, int value);
 /// @param list the linked list
 /// @param index the position in the list
 /// @return the value removed
-int ioopm_linked_list_remove(ioopm_list_t *list, int index);
+int ioopm_linked_list_remove(ioopm_list_t *list, size_t index);
 
 /// @brief Retrieve an element from a linked list in O(n) time.
 /// The valid values of index are [0,n-1] for a list of n elements,
@@ -48,7 +49,7 @@ int ioopm_linked_list_remove(ioopm_list_t *list, int index);
 /// @param list the linked list that will be extended
 /// @param index the position in the list
 /// @return the value at the given position
-int ioopm_linked_list_get(ioopm_list_t *list, int index);
+int ioopm_linked_list_get(ioopm_list_t *list, size_t index);
 
 /// @brief Test if an element is in the list
 /// @param list the linked list
@@ -59,7 +60,7 @@ bool ioopm_linked_list_contains(ioopm_list_t *list, int element);
 /// @brief Lookup the number of elements in the linked list in O(1) time
 /// @param list the linked list
 /// @return the number of elements in the list
-int ioopm_linked_list_size(ioopm_list_t *list);
+size_t ioopm_linked_list_size(ioopm_list_t *list);
 
 /// @brief Test whether a list is empty or not
 /// @param list the linked list
